fixed element_buffer warning

+++ pyro/src/platform/opengl/gl_vertex_array.cpp
@@ -60,7 +60,8 @@ void pyro::gl_vertex_array::add_buffer(const ref<vertex_buffer>& vertex_buffer)
             to_opengl_type(element.type), 
             element.normalised ? GL_TRUE : GL_FALSE, 
             layout.stride(), 
-            reinterpret_cast<const void*>((intptr_t)element.offset));
+            reinterpret_cast<const void*>(element.offset)
+        );
         m_vertex_buffer_index++;
     }
     m_vertex_buffers.push_back(vertex_buffer);
